+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to 
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+- Ports the new range slider for word boundaries to the iOS app (ABY-18).
+
 ## [2.0-13]
 - Introduces a brand-new redesign of the app on Mac with more information at a glance (ABY-11).